requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python
pillow
realesrgan
numpy==1.26.4
fastapi
uvicorn
realesrgan
opencv-python
numpy<2